test: address review feedback on survey deserializer test
- Rename to surveys_test.dart (not issue-specific)
- Use Object? instead of dynamic (satisfies avoid_annotating_with_dynamic)
- Replace the global _absent sentinel with an includeLink param
